Move achievement calculations to own file
